Math1f/target/clover/src-instrumented/org/apache/commons/math3/fraction/BigFractionFormat.java
/* $$ This file has been instrumented by Clover 4.4.1#2019101123313948 $$ *//*
 * Licensed to the Apache Software Foundation (ASF) under one or more
 * contributor license agreements.  See the NOTICE file distributed with
 * this work for additional information regarding copyright ownership.
 * The ASF licenses this file to You under the Apache License, Version 2.0
 * (the "License"); you may not use this file except in compliance with
 * the License.  You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package org.apache.commons.math3.fraction;

import java.io.Serializable;
import java.math.BigInteger;
import java.text.FieldPosition;
import java.text.NumberFormat;
import java.text.ParsePosition;
import java.util.Locale;

import org.apache.commons.math3.exception.MathIllegalArgumentException;
import org.apache.commons.math3.exception.MathParseException;
import org.apache.commons.math3.exception.util.LocalizedFormats;

/**
 * Formats a BigFraction number in proper format or improper format.
 * <p>
 * The number format for each of the whole number, numerator and,
 * denominator can be configured.
 * </p>
 *
 * @since 2.0
 * @version $Id$
 */
@java.lang.SuppressWarnings({"fallthrough"}) public class BigFractionFormat extends AbstractFormat implements Serializable {public static class __CLR4_4_1b92b92lb90p7vw{public static com_atlassian_clover.CoverageRecorder R;public static com_atlassian_clover.CloverProfile[] profiles = { };@java.lang.SuppressWarnings("unchecked") public static <I, T extends I> I lambdaInc(final int i,final T l,final int si){java.lang.reflect.InvocationHandler h=new java.lang.reflect.InvocationHandler(){public java.lang.Object invoke(java.lang.Object p,java.lang.reflect.Method m,java.lang.Object[] a) throws Throwable{R.inc(i);R.inc(si);try{return m.invoke(l,a);}catch(java.lang.reflect.InvocationTargetException e){throw e.getCause()!=null?e.getCause():new RuntimeException("Clover failed to invoke instrumented lambda",e);}}};return (I)java.lang.reflect.Proxy.newProxyInstance(l.getClass().getClassLoader(),l.getClass().getInterfaces(),h);}static{com_atlassian_clover.CoverageRecorder _R=null;try{com_atlassian_clover.CloverVersionInfo.An_old_version_of_clover_is_on_your_compilation_classpath___Please_remove___Required_version_is___4_4_1();if(2019101123313948L!=com_atlassian_clover.CloverVersionInfo.getBuildStamp()){com_atlassian_clover.Clover.l("[CLOVER] WARNING: The Clover version used in instrumentation does not match the runtime version. You need to run instrumented classes against the same version of Clover that you instrumented with.");com_atlassian_clover.Clover.l("[CLOVER] WARNING: Instr=4.4.1#2019101123313948,Runtime="+com_atlassian_clover.CloverVersionInfo.getReleaseNum()+"#"+com_atlassian_clover.CloverVersionInfo.getBuildStamp());}R=com_atlassian_clover.Clover.getNullRecorder();_R=com_atlassian_clover.Clover.getNullRecorder();_R=com_atlassian_clover.Clover.getRecorder("\u002f\u0068\u006f\u006d\u0065\u002f\u0073\u0065\u0070\u0065\u0068\u0072\u002f\u006d\u0061\u0073\u0074\u0065\u0072\u0073\u002f\u0074\u0065\u0073\u0074\u005f\u0067\u0065\u006e\u0065\u0072\u0061\u0074\u0069\u006f\u006e\u002f\u0061\u006c\u006c\u005f\u0070\u0072\u006f\u006a\u0065\u0063\u0074\u0073\u005f\u0063\u006f\u0076\u0065\u0072\u0061\u0067\u0065\u002f\u004d\u0061\u0074\u0068\u0031\u0066\u002f\u0074\u0061\u0072\u0067\u0065\u0074\u002f\u0063\u006c\u006f\u0076\u0065\u0072\u002f\u0063\u006c\u006f\u0076\u0065\u0072\u002e\u0064\u0062",1670137421344L,8589935092L,14673,profiles,new java.lang.String[]{"clover.distributed.coverage",null});}catch(java.lang.SecurityException e){java.lang.System.err.println("[CLOVER] FATAL ERROR: Clover could not be initialised because it has insufficient security privileges. Please consult the Clover documentation on the security policy file changes required. ("+e.getClass()+":"+e.getMessage()+")");}catch(java.lang.NoClassDefFoundError e){java.lang.System.err.println("[CLOVER] FATAL ERROR: Clover could not be initialised. Are you sure you have Clover in the runtime classpath? ("+e.getClass()+":"+e.getMessage()+")");}catch(java.lang.Throwable t){java.lang.System.err.println("[CLOVER] FATAL ERROR: Clover could not be initialised because of an unexpected error. ("+t.getClass()+":"+t.getMessage()+")");}R=_R;}}public static final com_atlassian_clover.TestNameSniffer __CLR4_4_1_TEST_NAME_SNIFFER=com_atlassian_clover.TestNameSniffer.NULL_INSTANCE;

    /** Serializable version identifier */
    private static final long serialVersionUID = -2932167925527338976L;

    /**
     * Create an improper formatting instance with the default number format
     * for the numerator and denominator.
     */
    public BigFractionFormat() {try{__CLR4_4_1b92b92lb90p7vw.R.inc(14582);
    }finally{__CLR4_4_1b92b92lb90p7vw.R.flushNeeded();}}

    /**
     * Create an improper formatting instance with a custom number format for
     * both the numerator and denominator.
     * @param format the custom format for both the numerator and denominator.
     */
    public BigFractionFormat(final NumberFormat format) {
        super(format);__CLR4_4_1b92b92lb90p7vw.R.inc(14584);try{__CLR4_4_1b92b92lb90p7vw.R.inc(14583);
    }finally{__CLR4_4_1b92b92lb90p7vw.R.flushNeeded();}}

    /**
     * Create an improper formatting instance with a custom number format for
     * the numerator and a custom number format for the denominator.
     * @param numeratorFormat the custom format for the numerator.
     * @param denominatorFormat the custom format for the denominator.
     */
    public BigFractionFormat(final NumberFormat numeratorFormat,
                             final NumberFormat denominatorFormat) {
        super(numeratorFormat, denominatorFormat);__CLR4_4_1b92b92lb90p7vw.R.inc(14586);try{__CLR4_4_1b92b92lb90p7vw.R.inc(14585);
    }finally{__CLR4_4_1b92b92lb90p7vw.R.flushNeeded();}}

    /**
     * Get the set of locales for which complex formats are available.  This
     * is the same set as the {@link NumberFormat} set.
     * @return available complex format locales.
     */
    public static Locale[] getAvailableLocales() {try{__CLR4_4_1b92b92lb90p7vw.R.inc(14587);
        __CLR4_4_1b92b92lb90p7vw.R.inc(14588);return NumberFormat.getAvailableLocales();
    }finally{__CLR4_4_1b92b92lb90p7vw.R.flushNeeded();}}

    /**
     * This static method calls formatBigFraction() on a default instance of
     * BigFractionFormat.
     *
     * @param f BigFraction object to format
     * @return A formatted BigFraction in proper form.
     */
    public static String formatBigFraction(final BigFraction f) {try{__CLR4_4_1b92b92lb90p7vw.R.inc(14589);
        __CLR4_4_1b92b92lb90p7vw.R.inc(14590);return getImproperInstance().format(f);
    }finally{__CLR4_4_1b92b92lb90p7vw.R.flushNeeded();}}

    /**
     * Returns the default complex format for the current locale.
     * @return the default complex format.
     */
    public static BigFractionFormat getImproperInstance() {try{__CLR4_4_1b92b92lb90p7vw.R.inc(14591);
        __CLR4_4_1b92b92lb90p7vw.R.inc(14592);return getImproperInstance(Locale.getDefault());
    }finally{__CLR4_4_1b92b92lb90p7vw.R.flushNeeded();}}

    /**
     * Returns the default complex format for the given locale.
     * @param locale the specific locale used by the format.
     * @return the complex format specific to the given locale.
     */
    public static BigFractionFormat getImproperInstance(final Locale locale) {try{__CLR4_4_1b92b92lb90p7vw.R.inc(14593);
        __CLR4_4_1b92b92lb90p7vw.R.inc(14594);return new BigFractionFormat(getDefaultNumberFormat(locale));
    }finally{__CLR4_4_1b92b92lb90p7vw.R.flushNeeded();}}

    /**
     * Returns the default complex format for the current locale.
     * @return the default complex format.
     */
    public static BigFractionFormat getProperInstance() {try{__CLR4_4_1b92b92lb90p7vw.R.inc(14595);
        __CLR4_4_1b92b92lb90p7vw.R.inc(14596);return getProperInstance(Locale.getDefault());
    }finally{__CLR4_4_1b92b92lb90p7vw.R.flushNeeded();}}

    /**
     * Returns the default complex format for the given locale.
     * @param locale the specific locale used by the format.
     * @return the complex format specific to the given locale.
     */
    public static BigFractionFormat getProperInstance(final Locale locale) {try{__CLR4_4_1b92b92lb90p7vw.R.inc(14597);
        __CLR4_4_1b92b92lb90p7vw.R.inc(14598);return new ProperBigFractionFormat(getDefaultNumberFormat(locale));
    }finally{__CLR4_4_1b92b92lb90p7vw.R.flushNeeded();}}

    /**
     * Formats a {@link BigFraction} object to produce a string.  The BigFraction is
     * output in improper format.
     *
     * @param BigFraction the object to format.
     * @param toAppendTo where the text is to be appended
     * @param pos On input: an alignment field, if desired. On output: the
     *            offsets of the alignment field
     * @return the value passed in as toAppendTo.
     */
    public StringBuffer format(final BigFraction BigFraction,
                               final StringBuffer toAppendTo, final FieldPosition pos) {try{__CLR4_4_1b92b92lb90p7vw.R.inc(14599);

        __CLR4_4_1b92b92lb90p7vw.R.inc(14600);pos.setBeginIndex(0);
        __CLR4_4_1b92b92lb90p7vw.R.inc(14601);pos.setEndIndex(0);

        __CLR4_4_1b92b92lb90p7vw.R.inc(14602);getNumeratorFormat().format(BigFraction.getNumerator(), toAppendTo, pos);
        __CLR4_4_1b92b92lb90p7vw.R.inc(14603);toAppendTo.append(" / ");
        __CLR4_4_1b92b92lb90p7vw.R.inc(14604);getDenominatorFormat().format(BigFraction.getDenominator(), toAppendTo, pos);

        __CLR4_4_1b92b92lb90p7vw.R.inc(14605);return toAppendTo;
    }finally{__CLR4_4_1b92b92lb90p7vw.R.flushNeeded();}}

    /**
     * Formats an object and appends the result to a StringBuffer.
     * <code>obj</code> must be either a  {@link BigFraction} object or a
     * {@link BigInteger} object or a {@link Number} object. Any other type of
     * object will result in an {@link IllegalArgumentException} being thrown.
     *
     * @param obj the object to format.
     * @param toAppendTo where the text is to be appended
     * @param pos On input: an alignment field, if desired. On output: the
     *            offsets of the alignment field
     * @return the value passed in as toAppendTo.
     * @see java.text.Format#format(java.lang.Object, java.lang.StringBuffer, java.text.FieldPosition)
     * @throws MathIllegalArgumentException if <code>obj</code> is not a valid type.
     */
    @Override
    public StringBuffer format(final Object obj,
                               final StringBuffer toAppendTo, final FieldPosition pos) {try{__CLR4_4_1b92b92lb90p7vw.R.inc(14606);

        __CLR4_4_1b92b92lb90p7vw.R.inc(14607);final StringBuffer ret;
        __CLR4_4_1b92b92lb90p7vw.R.inc(14608);if ((((obj instanceof BigFraction)&&(__CLR4_4_1b92b92lb90p7vw.R.iget(14609)!=0|true))||(__CLR4_4_1b92b92lb90p7vw.R.iget(14610)==0&false))) {{
            __CLR4_4_1b92b92lb90p7vw.R.inc(14611);ret = format((BigFraction) obj, toAppendTo, pos);
        } }else {__CLR4_4_1b92b92lb90p7vw.R.inc(14612);if ((((obj instanceof BigInteger)&&(__CLR4_4_1b92b92lb90p7vw.R.iget(14613)!=0|true))||(__CLR4_4_1b92b92lb90p7vw.R.iget(14614)==0&false))) {{
            __CLR4_4_1b92b92lb90p7vw.R.inc(14615);ret = format(new BigFraction((BigInteger) obj), toAppendTo, pos);
        } }else {__CLR4_4_1b92b92lb90p7vw.R.inc(14616);if ((((obj instanceof Number)&&(__CLR4_4_1b92b92lb90p7vw.R.iget(14617)!=0|true))||(__CLR4_4_1b92b92lb90p7vw.R.iget(14618)==0&false))) {{
            __CLR4_4_1b92b92lb90p7vw.R.inc(14619);ret = format(new BigFraction(((Number) obj).doubleValue()),
                         toAppendTo, pos);
        } }else {{
            __CLR4_4_1b92b92lb90p7vw.R.inc(14620);throw new MathIllegalArgumentException(LocalizedFormats.CANNOT_FORMAT_OBJECT_TO_FRACTION);
        }

        }}}__CLR4_4_1b92b92lb90p7vw.R.inc(14621);return ret;
    }finally{__CLR4_4_1b92b92lb90p7vw.R.flushNeeded();}}

    /**
     * Parses a string to produce a {@link BigFraction} object.
     * @param source the string to parse
     * @return the parsed {@link BigFraction} object.
     * @exception MathParseException if the beginning of the specified string
     *            cannot be parsed.
     */
    @Override
    public BigFraction parse(final String source) throws MathParseException {try{__CLR4_4_1b92b92lb90p7vw.R.inc(14622);
        __CLR4_4_1b92b92lb90p7vw.R.inc(14623);final ParsePosition parsePosition = new ParsePosition(0);
        __CLR4_4_1b92b92lb90p7vw.R.inc(14624);final BigFraction result = parse(source, parsePosition);
        __CLR4_4_1b92b92lb90p7vw.R.inc(14625);if ((((parsePosition.getIndex() == 0)&&(__CLR4_4_1b92b92lb90p7vw.R.iget(14626)!=0|true))||(__CLR4_4_1b92b92lb90p7vw.R.iget(14627)==0&false))) {{
            __CLR4_4_1b92b92lb90p7vw.R.inc(14628);throw new MathParseException(source, parsePosition.getErrorIndex(), BigFraction.class);
        }
        }__CLR4_4_1b92b92lb90p7vw.R.inc(14629);return result;
    }finally{__CLR4_4_1b92b92lb90p7vw.R.flushNeeded();}}

    /**
     * Parses a string to produce a {@link BigFraction} object.
     * This method expects the string to be formatted as an improper BigFraction.
     * @param source the string to parse
     * @param pos input/output parsing parameter.
     * @return the parsed {@link BigFraction} object.
     */
    @Override
    public BigFraction parse(final String source, final ParsePosition pos) {try{__CLR4_4_1b92b92lb90p7vw.R.inc(14630);
        __CLR4_4_1b92b92lb90p7vw.R.inc(14631);final int initialIndex = pos.getIndex();

        // parse whitespace
        __CLR4_4_1b92b92lb90p7vw.R.inc(14632);parseAndIgnoreWhitespace(source, pos);

        // parse numerator
        __CLR4_4_1b92b92lb90p7vw.R.inc(14633);final BigInteger num = parseNextBigInteger(source, pos);
        __CLR4_4_1b92b92lb90p7vw.R.inc(14634);if ((((num == null)&&(__CLR4_4_1b92b92lb90p7vw.R.iget(14635)!=0|true))||(__CLR4_4_1b92b92lb90p7vw.R.iget(14636)==0&false))) {{
            // invalid integer number
            // set index back to initial, error index should already be set
            // character examined.
            __CLR4_4_1b92b92lb90p7vw.R.inc(14637);pos.setIndex(initialIndex);
            __CLR4_4_1b92b92lb90p7vw.R.inc(14638);return null;
        }

        // parse '/'
        }__CLR4_4_1b92b92lb90p7vw.R.inc(14639);final int startIndex = pos.getIndex();
        __CLR4_4_1b92b92lb90p7vw.R.inc(14640);final char c = parseNextCharacter(source, pos);
        boolean __CLB4_4_1_bool0=false;__CLR4_4_1b92b92lb90p7vw.R.inc(14641);switch (c) {
        case 0 :if (!__CLB4_4_1_bool0) {__CLR4_4_1b92b92lb90p7vw.R.inc(14642);__CLB4_4_1_bool0=true;}
            // no '/'
            // return num as a BigFraction
            __CLR4_4_1b92b92lb90p7vw.R.inc(14643);return new BigFraction(num);
        case '/' :if (!__CLB4_4_1_bool0) {__CLR4_4_1b92b92lb90p7vw.R.inc(14644);__CLB4_4_1_bool0=true;}
            // found '/', continue parsing denominator
            __CLR4_4_1b92b92lb90p7vw.R.inc(14645);break;
        default :if (!__CLB4_4_1_bool0) {__CLR4_4_1b92b92lb90p7vw.R.inc(14646);__CLB4_4_1_bool0=true;}
            // invalid '/'
            // set index back to initial, error index should be the last
            // character examined.
            __CLR4_4_1b92b92lb90p7vw.R.inc(14647);pos.setIndex(initialIndex);
            __CLR4_4_1b92b92lb90p7vw.R.inc(14648);pos.setErrorIndex(startIndex);
            __CLR4_4_1b92b92lb90p7vw.R.inc(14649);return null;
        }

        // parse whitespace
        __CLR4_4_1b92b92lb90p7vw.R.inc(14650);parseAndIgnoreWhitespace(source, pos);

        // parse denominator
        __CLR4_4_1b92b92lb90p7vw.R.inc(14651);final BigInteger den = parseNextBigInteger(source, pos);
        __CLR4_4_1b92b92lb90p7vw.R.inc(14652);if ((((den == null)&&(__CLR4_4_1b92b92lb90p7vw.R.iget(14653)!=0|true))||(__CLR4_4_1b92b92lb90p7vw.R.iget(14654)==0&false))) {{
            // invalid integer number
            // set index back to initial, error index should already be set
            // character examined.
            __CLR4_4_1b92b92lb90p7vw.R.inc(14655);pos.setIndex(initialIndex);
            __CLR4_4_1b92b92lb90p7vw.R.inc(14656);return null;
        }

        }__CLR4_4_1b92b92lb90p7vw.R.inc(14657);return new BigFraction(num, den);
    }finally{__CLR4_4_1b92b92lb90p7vw.R.flushNeeded();}}

    /**
     * Parses a string to produce a <code>BigInteger</code>.
     * @param source the string to parse
     * @param pos input/output parsing parameter.
     * @return a parsed <code>BigInteger</code> or null if string does not
     * contain a BigInteger at the specified position
     */
    protected BigInteger parseNextBigInteger(final String source,
                                             final ParsePosition pos) {try{__CLR4_4_1b92b92lb90p7vw.R.inc(14658);

        __CLR4_4_1b92b92lb90p7vw.R.inc(14659);final int start = pos.getIndex();
         __CLR4_4_1b92b92lb90p7vw.R.inc(14660);int end = ((((source.charAt(start) == '-') )&&(__CLR4_4_1b92b92lb90p7vw.R.iget(14661)!=0|true))||(__CLR4_4_1b92b92lb90p7vw.R.iget(14662)==0&false))? (start + 1) : start;
         __CLR4_4_1b92b92lb90p7vw.R.inc(14663);while(((((end < source.length()) &&
               Character.isDigit(source.charAt(end)))&&(__CLR4_4_1b92b92lb90p7vw.R.iget(14664)!=0|true))||(__CLR4_4_1b92b92lb90p7vw.R.iget(14665)==0&false))) {{
             __CLR4_4_1b92b92lb90p7vw.R.inc(14666);++end;
         }

         }__CLR4_4_1b92b92lb90p7vw.R.inc(14667);try {
             __CLR4_4_1b92b92lb90p7vw.R.inc(14668);BigInteger n = new BigInteger(source.substring(start, end));
             __CLR4_4_1b92b92lb90p7vw.R.inc(14669);pos.setIndex(end);
             __CLR4_4_1b92b92lb90p7vw.R.inc(14670);return n;
         } catch (NumberFormatException nfe) {
             __CLR4_4_1b92b92lb90p7vw.R.inc(14671);pos.setErrorIndex(start);
             __CLR4_4_1b92b92lb90p7vw.R.inc(14672);return null;
         }

    }finally{__CLR4_4_1b92b92lb90p7vw.R.flushNeeded();}}

}
